requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dbt-core==1.7.1
dbt-postgres==1.7.1
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.0
neo4j==5.14.0
pandas==2.1.0 
numpy==1.24.3
apoc==1.0.0